# api/openapi.yaml
openapi: 3.1.0
info:
  title: Stadtfest-Finder API
  version: 0.1.0
  summary: REST API for the Stadtfest-Finder mobile app.
  description: |
    Single source of truth for the REST interface (API-first, see CLAUDE.md).
    Times are ISO 8601; "today" is evaluated in Europe/Berlin. Coordinates are WGS84.
    Public endpoints work without a token; protected ones require a JWT issued by the IdP.
  contact:
    name: Stadtfest-Finder
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /v1/events:
    get:
      operationId: searchEvents
      tags: [events]
      summary: Search events for map and list
      description: |
        Published or cancelled events that have not ended yet (end date >= today, Europe/Berlin).
        Sorted: running first, then by start date ascending, then by name.
        `lat`/`lon` define the reference point for `distanceKm` and, together with `radiusKm`,
        a radius filter. Without `lat`/`lon` the distance refers to the center of `bbox`.
      security: []
      parameters:
        - $ref: "#/components/parameters/Bbox"
        - $ref: "#/components/parameters/Lat"
        - $ref: "#/components/parameters/Lon"
        - $ref: "#/components/parameters/RadiusKm"
        - $ref: "#/components/parameters/When"
        - $ref: "#/components/parameters/Months"
        - $ref: "#/components/parameters/Categories"
        - $ref: "#/components/parameters/Query"
        - name: cursor
          in: query
          required: false
          description: >-
            Opaque cursor from `nextCursor` of the previous page. Cursors not issued by the
            server are rejected with `422`.
          schema:
            type: string
            maxLength: 512
        - name: limit
          in: query
          required: false
          description: Page size (the map requests up to 500).
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        "200":
          description: One page of events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "422":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /v1/events/count:
    get:
      operationId: countEvents
      tags: [events]
      summary: Count matching events
      description: |
        Number of events matching the same filters as `searchEvents`, used for the filter
        button and the chip counts. `byCategory` counts, per category, the events that match
        all filters except the category filter.
      security: []
      parameters:
        - $ref: "#/components/parameters/Bbox"
        - $ref: "#/components/parameters/Lat"
        - $ref: "#/components/parameters/Lon"
        - $ref: "#/components/parameters/RadiusKm"
        - $ref: "#/components/parameters/When"
        - $ref: "#/components/parameters/Months"
        - $ref: "#/components/parameters/Categories"
        - $ref: "#/components/parameters/Query"
      responses:
        "200":
          description: Counts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCount"
        "422":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /v1/events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Event detail
      description: |
        Full details of a published or cancelled event. Past events remain available by ID
        (timeline, lists). Drafts and deleted events yield `404`. With a token the response
        additionally contains `isFavorite`; such responses are never cached.
      security: []
      parameters:
        - name: eventId
          in: path
          required: true
          description: Event ID.
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Lat"
        - $ref: "#/components/parameters/Lon"
      responses:
        "200":
          description: Event detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/mod/events:
    get:
      operationId: listModEvents
      tags: [moderation]
      summary: Events of the moderator's region
      description: |
        All events of the caller's region except deleted ones, with favorite count, source and
        version. Ordered: upcoming by start date, then past ones descending, then events
        without date. `ids` restricts the list; IDs of other regions are silently dropped.
      parameters:
        - name: status
          in: query
          required: false
          description: Only events with this (derived) status.
          schema:
            $ref: "#/components/schemas/ModEventStatus"
        - name: q
          in: query
          required: false
          description: Search in name, place and city.
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: ids
          in: query
          required: false
          description: Comma-separated event IDs.
          style: form
          explode: false
          schema:
            type: array
            maxItems: 100
            items:
              type: string
              format: uuid
      responses:
        "200":
          description: Events of the region.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModEventList"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createModEvent
      tags: [moderation]
      summary: Create an event
      description: Always creates a draft in the caller's region. Only the name is required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModEventCreate"
      responses:
        "201":
          description: The created draft.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModEventDetail"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /v1/mod/events/{eventId}:
    parameters:
      - $ref: "#/components/parameters/ModEventId"
    get:
      operationId: getModEvent
      tags: [moderation]
      summary: Event for editing
      description: All fields of an event of the caller's region, with `version` and `ETag`.
      responses:
        "200":
          description: All fields and the version.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModEventDetail"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateModEvent
      tags: [moderation]
      summary: Edit an event
      description: |
        JSON merge patch: fields that are present are set (`null` clears them), the program is
        replaced as a whole. `If-Match` must carry the current version; otherwise `409
        version_conflict`. Cancelled events only accept text fields (`409
        invalid_transition`). Changes to a published event emit `event.updated`.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/ModEventPatch"
      responses:
        "200":
          description: The updated event.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModEventDetail"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "428":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteModEvent
      tags: [moderation]
      summary: Delete an event
      description: |
        Soft delete from any status. Favorites of the event are removed in the background;
        users are not notified (cancel instead to inform them).
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
  headers:
    ETag:
      description: Current version of the event, quoted (e.g. `"3"`); send it as `If-Match`.
      schema:
        type: string
  parameters:
...
    ModEventId:
      name: eventId
      in: path
      required: true
      description: Event ID.
      schema:
        type: string
        format: uuid
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Version from `ETag` / `version` (quoted or plain). Missing → `428`.
      schema:
        type: string
        minLength: 1
        maxLength: 20
    Bbox:
      name: bbox
      in: query
      required: false
      description: Map viewport `minLon,minLat,maxLon,maxLat`.
      style: form
      explode: false
      schema:
        type: array
        minItems: 4
        maxItems: 4
        items:
          type: number
    Lat:
      name: lat
      in: query
      required: false
      description: Latitude of the reference point (user position or map center). Requires `lon`.
      schema:
        type: number
        minimum: -90
        maximum: 90
    Lon:
      name: lon
      in: query
      required: false
      description: Longitude of the reference point. Requires `lat`.
      schema:
        type: number
        minimum: -180
        maximum: 180
    RadiusKm:
      name: radiusKm
      in: query
      required: false
      description: Radius around `lat`/`lon` in km. Ignored without `lat`/`lon`.
      schema:
        type: integer
        minimum: 10
        maximum: 300
        default: 150
    When:
      name: when
      in: query
      required: false
      description: |
        Time filter: `today` = running today; `weekend` = overlaps Saturday or Sunday of the
        current week; `months` = overlaps at least one month in `months`.
      schema:
        type: string
        enum: [all, today, weekend, months]
        default: all
    Months:
      name: months
      in: query
      required: false
      description: Months `YYYY-MM` (required for `when=months`, at most 12).
      style: form
      explode: false
      schema:
        type: array
        minItems: 1
        maxItems: 12
        items:
          type: string
          pattern: "^20[0-9]{2}-(0[1-9]|1[0-2])$"
    Categories:
      name: categories
      in: query
      required: false
      description: Category IDs. Unknown or inactive IDs are ignored silently.
      style: form
      explode: false
      schema:
        type: array
        maxItems: 50
        items:
          type: string
          format: uuid
    Query:
      name: q
      in: query
      required: false
      description: Text search in name, place and city (case/accent-insensitive, typo-tolerant).
      schema:
        type: string
        minLength: 2
        maxLength: 100
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by the IdP (Zitadel in production, Keycloak locally).
  responses:
    Error:
      description: Error in the common error format.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    CategoryRef:
      type: object
      description: Category as embedded in an event detail.
      required: [id, name, emoji, color]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        emoji:
          type: string
        color:
          type: string
          pattern: "^#[0-9A-F]{6}$"
    PublicEventStatus:
      type: string
      description: Status of a publicly visible event. "Past" is derived from `endDate`.
      enum: [published, cancelled]
    Image:
      type: object
      description: |
        Event image variants (R08), WebP with immutable URLs: `url` (long edge 1,600 px),
        `cardUrl` (600 px), `thumbUrl` (200 px). `jpegUrl` is the full size as JPEG, e.g. for
        link previews.
      required: [url, thumbUrl]
      properties:
        url:
          type: string
          format: uri
        cardUrl:
          type: string
          format: uri
        thumbUrl:
          type: string
          format: uri
        jpegUrl:
          type: string
          format: uri
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
    EventSummary:
      type: object
      description: Event as shown in carousel, list and timeline.
      required: [id, name, shortName, status, startDate, endDate, place, city, lat, lon, categoryId]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        shortName:
          type: string
          description: Short name for the selected map pin.
        status:
          $ref: "#/components/schemas/PublicEventStatus"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        place:
          type: string
        city:
          type: string
        lat:
          type: number
        lon:
          type: number
        categoryId:
          type: string
          format: uuid
        distanceKm:
          type: ["number", "null"]
          description: Distance to the reference point, if one was given.
        coverImage:
          oneOf:
            - $ref: "#/components/schemas/Image"
            - type: "null"
    EventPage:
      type: object
      description: One page of events.
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventSummary"
        nextCursor:
          type: ["string", "null"]
          description: Cursor for the next page, `null` on the last page.
    EventCount:
      type: object
      description: Result counts for filter previews.
      required: [total, byCategory]
      properties:
        total:
          type: integer
          minimum: 0
        byCategory:
          type: object
          description: Category ID -> number of events matching all filters except categories.
          additionalProperties:
            type: integer
            minimum: 0
    ProgramItem:
      type: object
      description: Entry of the event program.
      required: [date, timeLabel, title]
      properties:
        date:
          type: string
          format: date
        timeLabel:
          type: string
          examples: ["10:45 Uhr"]
        title:
          type: string
        subtitle:
          type: ["string", "null"]
    EventDetail:
      type: object
      description: Full public event detail.
      required:
        [id, name, shortName, status, category, startDate, endDate, openingHours, place, address,
         city, postalCode, lat, lon, program, images]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        shortName:
          type: string
        status:
          $ref: "#/components/schemas/PublicEventStatus"
        cancelReason:
          type: ["string", "null"]
        category:
          $ref: "#/components/schemas/CategoryRef"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        openingHours:
          type: array
          items:
            type: string
        price:
          type: ["string", "null"]
        place:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
          pattern: "^[0-9]{5}$"
        lat:
          type: number
        lon:
          type: number
        description:
          type: ["string", "null"]
        program:
          type: array
          items:
            $ref: "#/components/schemas/ProgramItem"
        transit:
          type: ["string", "null"]
        parking:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
          format: uri
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        distanceKm:
          type: ["number", "null"]
        isFavorite:
          type: boolean
          description: Whether the caller marked the event as favorite. Only present with a token.
...
    ModEventStatus:
      type: string
      description: |
        Status in the moderation view. `past` is derived for published events whose end date
        is before today (Europe/Berlin).
      enum: [draft, published, past, cancelled]
    EventSource:
      type: string
      description: "`manual` or found by the AI search (`ai`, R10)."
      enum: [manual, ai]
    ModEventSummary:
      type: object
      description: Row of the moderation overview.
      required: [id, name, status, place, city, favoriteCount, source, version]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: "#/components/schemas/ModEventStatus"
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        place:
          type: string
        city:
          type: string
        categoryId:
          type: ["string", "null"]
          format: uuid
        favoriteCount:
          type: integer
          minimum: 0
        source:
          $ref: "#/components/schemas/EventSource"
        version:
          type: integer
          minimum: 1
    ModEventList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModEventSummary"
    ModProgramItem:
      type: object
      description: Program entry as entered by the moderator.
      required: [date, timeLabel, title]
      properties:
        date:
          type: string
          format: date
        timeLabel:
          type: string
          maxLength: 40
        title:
          type: string
          minLength: 1
          maxLength: 120
        subtitle:
          type: ["string", "null"]
          maxLength: 200
    ModEventDetail:
      type: object
      description: Event with all editable fields; fields of drafts may be empty.
      required:
        [id, regionId, name, shortName, status, openingHours, place, address, city, program,
         favoriteCount, source, version, images]
      properties:
        id:
          type: string
          format: uuid
        regionId:
          type: string
          format: uuid
        name:
          type: string
        shortName:
          type: string
        status:
          $ref: "#/components/schemas/ModEventStatus"
        categoryId:
          type: ["string", "null"]
          format: uuid
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        openingHours:
          type: array
          items:
            type: string
        price:
          type: ["string", "null"]
        place:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lon:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        program:
          type: array
          items:
            $ref: "#/components/schemas/ModProgramItem"
        transit:
          type: ["string", "null"]
        parking:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
          format: date-time
        favoriteCount:
          type: integer
          minimum: 0
        source:
          $ref: "#/components/schemas/EventSource"
        version:
          type: integer
          minimum: 1
        images:
          type: array
          description: All images in order, including processing and failed ones (R08).
          items:
            $ref: "#/components/schemas/ModImage"
    ModImageStatus:
      type: string
      enum: [processing, ready, failed]
    ModImage:
      type: object
      description: An event image in the moderation view. Variants only when `ready`.
      required: [id, status, position, image]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ModImageStatus"
        position:
          type: integer
          minimum: 0
          description: 0 is the cover image.
        image:
          description: Variants; `null` until the image is ready.
          oneOf:
            - $ref: "#/components/schemas/Image"
            - type: "null"
...
    ModEventFields:
      type: object
      description: |
        Editable fields. Lengths: name 120, short name 18, description 5,000, 10 opening hour
        lines of 80, 50 program entries. Postal code and city come from geocoding.
      properties:
        shortName:
          type: ["string", "null"]
          maxLength: 18
          description: Defaults to the name shortened to 18 characters.
        categoryId:
          type: ["string", "null"]
          format: uuid
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        openingHours:
          type: ["array", "null"]
          maxItems: 10
          items:
            type: string
            maxLength: 80
        price:
          type: ["string", "null"]
          maxLength: 200
        place:
          type: ["string", "null"]
          maxLength: 120
        address:
          type: ["string", "null"]
          maxLength: 200
        city:
          type: ["string", "null"]
          maxLength: 120
        postalCode:
          type: ["string", "null"]
          pattern: "^[0-9]{5}$"
        lat:
          type: ["number", "null"]
          minimum: -90
          maximum: 90
        lon:
          type: ["number", "null"]
          minimum: -180
          maximum: 180
        description:
          type: ["string", "null"]
          maxLength: 5000
        program:
          type: ["array", "null"]
          maxItems: 50
          items:
            $ref: "#/components/schemas/ModProgramItem"
        transit:
          type: ["string", "null"]
          maxLength: 500
        parking:
          type: ["string", "null"]
          maxLength: 500
        websiteUrl:
          type: ["string", "null"]
          maxLength: 500
        cancelReason:
          type: ["string", "null"]
          maxLength: 500
    ModEventCreate:
      description: New draft; only the name is required.
      allOf:
        - $ref: "#/components/schemas/ModEventFields"
        - type: object
          required: [name]
          properties:
            name:
              type: string
              minLength: 1
              maxLength: 120
    ModEventPatch:
      description: Merge patch of an event; all fields optional.
      allOf:
        - $ref: "#/components/schemas/ModEventFields"
        - type: object
          properties:
            name:
              type: string
              minLength: 1
              maxLength: 120
...
    Error:
      type: object
      description: Common error format for all non-2xx responses.
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code, e.g. `validation_failed`, `not_found`.
          examples: [validation_failed]
        message:
          type: string
          description: Human-readable message (German, suitable for display).
          examples: [Bitte fülle die markierten Pflichtfelder aus]
        fields:
          type: object
          description: Field-level errors keyed by field name.
          additionalProperties:
            type: string
...
security:
  - bearerAuth: []